# producer/api/v1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for News Api",
        "title": "News Api",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/news": {
            "get": {
                "description": "Get the list of news",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "List news",
                "operationId": "list-news",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "rows",
                        "name": "rows",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.News"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new news",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Create a new news",
                "operationId": "create-new",
                "parameters": [
                    {
                        "description": "Create news",
                        "name": "new",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.News"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.News"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.News": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "created": {
                    "type": "string",
                    "example": "2020-04-21T00:00:00Z"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}